creating protopath
